+++ dsl/module-dsl/blocking/src/main/kotlin/se/occurrent/dsl/module/blocking/Module.kt
@@ -17,9 +17,10 @@
 package se.occurrent.dsl.module.blocking
 
 import org.occurrent.application.converter.CloudEventConverter
+import org.occurrent.application.typemapper.ReflectionTypeMapper
+import org.occurrent.application.typemapper.TypeMapper
 import org.occurrent.dsl.subscription.blocking.Subscriptions
 import org.occurrent.subscription.api.blocking.SubscriptionModel
-import kotlin.reflect.KClass
 
 /**
  * DSL marker annotation which is used to limit callers so that they will not have implicit access to multiple receivers whose classes are in the set of annotated classes.
@@ -29,10 +30,10 @@ import kotlin.reflect.KClass
 internal annotation class ModuleDSL
 
 fun <C : Any, E : Any> module(
-    cloudEventConverter: CloudEventConverter<E>, eventNameFromType: (KClass<out E>) -> String = { e -> e.simpleName!! },
+    cloudEventConverter: CloudEventConverter<E>, typeMapper: TypeMapper<E> = ReflectionTypeMapper.qualified(),
     b: (@ModuleDSL ModuleBuilder<C, E>).() -> Unit
 ): Module<C> {
-    val module = ModuleBuilder<C, E>(cloudEventConverter, eventNameFromType).apply(b)
+    val module = ModuleBuilder<C, E>(cloudEventConverter, typeMapper).apply(b)
 
     return object : Module<C> {
         override fun dispatch(vararg commands: C) {
@@ -44,7 +45,7 @@ fun <C : Any, E : Any> module(
 }
 
 @ModuleDSL
-class ModuleBuilder<C : Any, E : Any> internal constructor(private val cloudEventConverter: CloudEventConverter<E>, private val eventNameFromType: (KClass<out E>) -> String) {
+class ModuleBuilder<C : Any, E : Any> internal constructor(private val cloudEventConverter: CloudEventConverter<E>, private val typeMapper: TypeMapper<E>) {
     internal val commandDispatchers = mutableListOf<CommandDispatcher<C, out Any>>()
 
     fun <B : Any> commands(commandDispatcher: CommandDispatcher<C, B>, commands: (@ModuleDSL B).() -> Unit) {
@@ -57,7 +58,7 @@ class ModuleBuilder<C : Any, E : Any> internal constructor(private val cloudEven
     }
 
     fun subscriptions(subscriptionModel: SubscriptionModel, subscriptions: (@ModuleDSL Subscriptions<E>).() -> Unit) {
-        Subscriptions(subscriptionModel, cloudEventConverter, eventNameFromType).apply(subscriptions)
+        Subscriptions(subscriptionModel, cloudEventConverter, typeMapper).apply(subscriptions)
     }
 }