+++ dsl/module-dsl/blocking/src/test/kotlin/se/occurrent/dsl/module/blocking/MultipleCommandDispatchersTest.kt
@@ -54,7 +54,7 @@ class MultipleCommandDispatchersTest {
         val allEvents = CopyOnWriteArrayList<DomainEvent>()
 
         // Module Configuration
-        val module = module<Command, DomainEvent>(cloudEventConverter, eventNameFromType = { e -> e.qualifiedName!! }) {
+        val module = module<Command, DomainEvent>(cloudEventConverter) {
             commands(dispatchTo(applicationService)) {
                 command(DefineName::getId, Name::defineNameFromCommand)
             }