+++ dsl/query-dsl/blocking/src/main/java/org/occurrent/dsl/query/blocking/DomainEventQueries.java
@@ -18,13 +18,16 @@
 package org.occurrent.dsl.query.blocking;
 
 import io.cloudevents.CloudEvent;
+import org.jetbrains.annotations.Nullable;
 import org.occurrent.application.converter.CloudEventConverter;
+import org.occurrent.application.typemapper.ReflectionTypeMapper;
+import org.occurrent.application.typemapper.TypeMapper;
 import org.occurrent.eventstore.api.SortBy;
 import org.occurrent.eventstore.api.blocking.EventStoreQueries;
 import org.occurrent.filter.Filter;
 
-import java.util.Objects;
-import java.util.Optional;
+import java.util.*;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 /**
@@ -37,10 +40,17 @@ public class DomainEventQueries<T> {
 
     private final EventStoreQueries eventStoreQueries;
     private final CloudEventConverter<T> cloudEventConverter;
+    private final TypeMapper<T> typeMapper;
 
     public DomainEventQueries(EventStoreQueries eventStoreQueries, CloudEventConverter<T> cloudEventConverter) {
+        this(eventStoreQueries, cloudEventConverter, ReflectionTypeMapper.qualified());
+    }
+
+    public DomainEventQueries(EventStoreQueries eventStoreQueries, CloudEventConverter<T> cloudEventConverter, TypeMapper<T> typeMapper) {
+        this.typeMapper = typeMapper;
         Objects.requireNonNull(eventStoreQueries, EventStoreQueries.class.getSimpleName() + " cannot be null");
         Objects.requireNonNull(cloudEventConverter, CloudEventConverter.class.getSimpleName() + " cannot be null");
+        Objects.requireNonNull(typeMapper, TypeMapper.class.getSimpleName() + " cannot be null");
         this.eventStoreQueries = eventStoreQueries;
         this.cloudEventConverter = cloudEventConverter;
     }
@@ -55,12 +65,147 @@ public class DomainEventQueries<T> {
     }
 
     /**
+     * Query for the first event of the given type.
+     *
+     * @return All cloud events matching the specified type.
+     */
+    public <E extends T> Optional<E> queryOne(Class<E> type) {
+        return query(type).findFirst();
+    }
+
+    /**
+     * Query by event type (will use the supplied {@link TypeMapper} to get the cloud event type from the class.
+     * <p>
+     * Note that it's recommended to create an index the fields you're sorting on in order to make them efficient.
+     *
+     * @return All cloud events matching the specified type.
+     */
+    public <E extends T> Stream<E> query(Class<E> type) {
+        return this.toDomainEvents(eventStoreQueries.query(Filter.type(typeMapper.getCloudEventType(type))));
+    }
+
+    /**
+     * Query by event type (will use the supplied {@link TypeMapper} to get the cloud event type from the class, also include {@code skip} and {@code limit}.
+     * <p>
+     * Note that it's recommended to create an index the fields you're sorting on in order to make them efficient.
+     *
+     * @return All cloud events matching the specified type, skip and limit.
+     */
+    public <E extends T> Stream<E> query(Class<E> type, int skip, int limit) {
+        return this.toDomainEvents(eventStoreQueries.query(Filter.type(typeMapper.getCloudEventType(type)), skip, limit));
+    }
+
+    /**
+     * Query by event type (will use the supplied {@link TypeMapper} to get the cloud event type from the class, also include {@code skip}, {@code limit} and {@code sortBy}.
+     * <p>
+     * Note that it's recommended to create an index the fields you're sorting on in order to make them efficient.
+     *
+     * @return All cloud events matching the specified type, skip, limit and sort by <code>sortBy</code>.
+     */
+    public <E extends T> Stream<E> query(Class<E> type, int skip, int limit, SortBy sortBy) {
+        return this.toDomainEvents(eventStoreQueries.query(Filter.type(typeMapper.getCloudEventType(type)), skip, limit, sortBy));
+    }
+
+    /**
+     * Query by event type (will use the supplied {@link TypeMapper} to get the cloud event type from the class, also including sorting .
+     * <p>
+     * Note that it's recommended to create an index the fields you're sorting on in order to make them efficient.
+     *
+     * @return All cloud events matching the specified type, sorted by <code>sortBy</code>.
+     */
+    public <E extends T> Stream<E> query(Class<E> type, SortBy sortBy) {
+        return this.toDomainEvents(eventStoreQueries.query(Filter.type(typeMapper.getCloudEventType(type)), sortBy));
+    }
+
+    /**
      * Note that it's recommended to create an index the fields you're sorting on in order to make them efficient.
      *
      * @return All cloud events matching the specified filter, skip, limit and sort by <code>sortBy</code>.
      */
     public <E extends T> Stream<E> query(Filter filter, int skip, int limit, SortBy sortBy) {
         return toDomainEvents(eventStoreQueries.query(filter, skip, limit, sortBy));
+    }
+
+    /**
+     * Query by event types (will use the supplied {@link TypeMapper} to get the cloud event type from the class, also include {@code skip}, {@code limit} and {@code sortBy}.
+     * <p>
+     * Note that it's recommended to create an index the fields you're sorting on in order to make them efficient.
+     *
+     * @return All cloud events matching the specified type, skip, limit and sort by <code>sortBy</code>.
+     */
+    public Stream<T> query(Collection<Class<? extends T>> types, int skip, int limit, SortBy sortBy) {
+        final Filter filterToUse = createFilterFrom(types);
+        if (filterToUse == null) {
+            return Stream.empty();
+        }
+
+        return this.toDomainEvents(eventStoreQueries.query(filterToUse, skip, limit, sortBy));
+    }
+
+    /**
+     * Query by event types (will use the supplied {@link TypeMapper} to get the cloud event type from the class, also include {@code skip}, {@code limit}.
+     * <p>
+     * Note that it's recommended to create an index the fields you're sorting on in order to make them efficient.
+     *
+     * @return All cloud events matching the specified type, skip, limit.
+     */
+    public Stream<T> query(Collection<Class<? extends T>> types, int skip, int limit) {
+        final Filter filterToUse = createFilterFrom(types);
+        if (filterToUse == null) {
+            return Stream.empty();
+        }
+
+        return this.toDomainEvents(eventStoreQueries.query(filterToUse, skip, limit));
+    }
+
+    /**
+     * Query by event types (will use the supplied {@link TypeMapper} to get the cloud event type from the class, also include {@code sortBy}.
+     * <p>
+     * Note that it's recommended to create an index the fields you're sorting on in order to make them efficient.
+     *
+     * @return All cloud events matching the specified type, skip, limit.
+     */
+    public Stream<T> query(Collection<Class<? extends T>> types, SortBy sortBy) {
+        final Filter filterToUse = createFilterFrom(types);
+        if (filterToUse == null) {
+            return Stream.empty();
+        }
+
+        return this.toDomainEvents(eventStoreQueries.query(filterToUse, sortBy));
+    }
+
+    /**
+     * Query by event types (will use the supplied {@link TypeMapper} to get the cloud event type from the class..
+     * <p>
+     * Note that it's recommended to create an index the fields you're sorting on in order to make them efficient.
+     *
+     * @return All cloud events matching the specified type, skip, limit.
+     */
+    public Stream<T> query(Collection<Class<? extends T>> types) {
+        final Filter filterToUse = createFilterFrom(types);
+        if (filterToUse == null) {
+            return Stream.empty();
+        }
+
+        return this.toDomainEvents(eventStoreQueries.query(filterToUse));
+    }
+
+    /**
+     * Query by event types (will use the supplied {@link TypeMapper} to get the cloud event type from the class, also include {@code sortBy}.
+     * <p>
+     * Note that it's recommended to create an index the fields you're sorting on in order to make them efficient.
+     *
+     * @return All cloud events matching the specified type, skip, limit.
+     */
+    @SuppressWarnings("unchecked")
+    @SafeVarargs
+    public final Stream<T> query(Class<? extends T> type, Class<? extends T>... types) {
+        final List<T> list = new ArrayList<>();
+        list.add((T) type);
+        if (types != null && types.length > 0) {
+            list.addAll(Arrays.stream(types).map(t -> (T) t).collect(Collectors.toList()));
+        }
+        return query((Collection<Class<? extends T>>) list);
     }
 
     /**
@@ -142,5 +287,13 @@ public class DomainEventQueries<T> {
     @SuppressWarnings("unchecked")
     private <E extends T> Stream<E> toDomainEvents(Stream<CloudEvent> stream) {
         return stream.map(cloudEventConverter::toDomainEvent).map(t -> (E) t);
+    }
+
+    @Nullable
+    private Filter createFilterFrom(Collection<Class<? extends T>> types) {
+        return (types == null ? Stream.<Class<? extends T>>empty() : types.stream())
+                .map(type -> Filter.type(typeMapper.getCloudEventType(type)))
+                .reduce(Filter::or)
+                .orElse(null);
     }
 }