+++ changelog.md
@@ -1,5 +1,36 @@
 ### Changelog next version
 
+* Introduced the concept of TypeMapper's. A type mapper is component whose purpose it is to get the [cloud event type](https://occurrent.org/documentation#cloudevents) from your domain events.
+  Type mappers are now used by multiple modules in the Occurrent ecosystem, for example the [subscription dsl](https://occurrent.org/documentation#subscription-dsl), [cloud event mapper](https://occurrent.org/documentation#cloudevent-conversion),
+  and the new domain queries DSL. You should use the same type mapper instance for all these components. To write a custom type mapper, depend on the `org.occurent:type-mapper-api` module and implement the `org.occurrent.application.typemapper.TypeMapper`
+  (functional) interface.
+* Introduced a blocking Query DSL. It's a small wrapper around the [EventStoreQueries](https://occurrent.org/documentation#eventstore-queries) API that let's you work with domain events instead of CloudEvents. 
+  Depend on the `org.occurrent:query-dsl-blocking` module and create an instance of `org.occurrent.dsl.query.blocking.DomainEventQueries`. For example:
+
+  ```java                                                      
+  EventStoreQueries eventStoreQueries = .. 
+  CloudEventConverter<DomainEvent> cloudEventConverter = ..
+  TypeMapper<DomainEvent> typeMapper = ..
+  DomainEventQueries<DomainEvent> domainEventQueries = new DomainEventQueries<DomainEvent>(eventStoreQueries, cloudEventConverter, typeMapper);
+   
+  Stream<DomainEvent> events = domainQueries.query(Filter.subject("someSubject"));
+  ```
+  
+  There's also support for skip, limits and sorting and convenience methods for querying for a single event:
+
+  ```java                                                      
+  Stream<DomainEvent> events = domainQueries.query(GameStarted.class, GameEnded.class); // Find only events of this type
+  GameStarted event1 = domainQueries.queryOne(GameStarted.class); // Find the first event of this type
+  GamePlayed event2s = domainQueries.queryOne(Filter.id("d7542cef-ac20-4e74-9128-fdec94540fda")); // Find event with this id
+  ```
+  
+  There are also some Kotlin extensions that you can use to query for a `Sequence` of events instead of a `Stream`:
+
+  ```kotlin
+  val events : Sequence<DomainEvent> = domainQueries.queryForSequence(GamePlayed::class, GameWon::class, skip = 2) // Find only events of this type and skip the first two events
+  val event1 : GameStarted = domainQueries.querySingle() // Find the first event of this type
+  val event2 : GamePlayed = domainQueries.querySingle(Filter.id("d7542cef-ac20-4e74-9128-fdec94540fda")) // Find event with this id
+  ```
 * Introducing spring boot starter project to easily bootstrap Occurrent if using Spring. Depend on `org.occurrent:spring-boot-starter-mongodb` and create a Spring Boot application annotated with `@SpringBootApplication` as you would normally do.
   Occurrent will then configure the following components automatically:
     * Spring MongoDB Event Store instance (`EventStore`)
@@ -8,8 +39,8 @@
     * A Jackson-based `CloudEventConverter`
     * A `GenericApplication` instance (`ApplicationService`)
     * A subscription dsl instance (`Subscriptions`)
+    * A reflection based type mapper that uses the fully-qualified class name as cloud event type (you _should_ absolutely override this bean for production use cases) (`TypeMapper`)
   See `org.occurrent.springboot.OccurrentMongoAutoConfiguration` if you want to know exactly what gets configured.
-* Introduced `org.occurrent.dsl.subscription.blocking.EventNameFromType` is may be supplied to an instance of `org.occurrent.dsl.subscription.blocking.Subscriptions` that allows it to generate correct queries when subscribing.  
 * Upgraded spring-boot from 2.5.4 to 2.5.6.
 
 ## Changelog 0.13.1 (2021-10-03)